+++ src/app.ts
@@ -2,7 +2,7 @@ import express from 'express'
 import cors from 'cors'
 import favicon from 'express-favicon'
 import logger from 'morgan'
-import userRouter from './routes/user'
+import userRouter from './routes/auth'
 import plansRouter from './routes/plans'
 import accountRouter from './routes/account'
 import swaggerUI from 'swagger-ui-express'
